backend/scheduler-service/requirements.txt
APScheduler==3.11.0
kafka-python==2.0.2
python-dotenv==1.0.1
tzlocal==5.2
